Remove mangling
Mangling was added in Nov 16 2022 to make startup time faster

In reality, it was added to prevent extensions from modifying VSCode's codebase

This could make sense from a security standpoint, but in reality it's stupid. Extensions can run any executable or javascript already.

In my opinion, it would make more sense to invest the time of mangling an open source project into fixing the security issues with extensions.